Chore/parallel executor benches (#3045)
## Linked Issues/PRs
<!-- List of related issues/PRs -->

## Description
<!-- List of detailed changes -->
This in practice ended up being as much of an integration tests as a
benchmark. There were a few places that weren't plugged in that needed
finessing before this would run properly.

~In addition, this uses the `u32-tx-pointer` feature from `fuel-tx`.
This is a **_BREAKING_** change as the header has changed.~

~If we don't want to contaminate the code with a breaking change for the
benchmark, then we can hide that behind a new feature flag on the
`fuel-types` library.~

I've introduced a `u32-tx-count` feature that changes a lot of `u16` to
`u32` for tx pointers as well as tx count in header etc. This prevents
this code from being breaking.

## Checklist
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

+++ crates/services/txpool_v2/src/selection_algorithms/ratio_tip_gas.rs
@@ -152,7 +152,7 @@ where
         storage: &mut S,
     ) -> RemovedTransactions {
         let mut gas_left = constraints.max_gas;
-        let mut space_left = constraints.maximum_block_size as usize;
+        let mut space_left = constraints.maximum_block_size;
         let mut nb_left = constraints.maximum_txs;
         let mut result = Vec::new();
 
@@ -211,8 +211,9 @@ where
                 }
 
                 let not_enough_gas = stored_transaction.transaction.max_gas() > gas_left;
-                let too_big_tx =
-                    stored_transaction.transaction.metered_bytes_size() > space_left;
+                let too_big_tx = stored_transaction.transaction.metered_bytes_size()
+                    as u64
+                    > space_left;
 
                 if not_enough_gas || too_big_tx {
                     continue;
@@ -220,8 +221,9 @@ where
 
                 gas_left =
                     gas_left.saturating_sub(stored_transaction.transaction.max_gas());
-                space_left = space_left
-                    .saturating_sub(stored_transaction.transaction.metered_bytes_size());
+                space_left = space_left.saturating_sub(
+                    stored_transaction.transaction.metered_bytes_size() as u64,
+                );
                 nb_left = nb_left.saturating_sub(1);
 
                 let dependents = storage.get_dependents(storage_id).collect::<Vec<_>>();
